requirements_win.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
abseil-cpp=20210324.2=hd77b12b_0
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py39h2bbff1b_1
aiosignal=1.2.0=pyhd3eb1b0_0
astor=0.8.1=py39haa95532_0
astunparse=1.6.3=py_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py39haa95532_0
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2022.3.29=haa95532_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py39haa95532_2
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39haa95532_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.8=py39h71e12ea_0
dataclasses=0.8=pyh6d0b6a4_7
flask=2.0.3=pyhd3eb1b0_0
flatbuffers=2.0.0=h6c2663c_0
freeglut=3.2.2=h0e60522_1
freetype=2.10.4=h546665d_1
frozenlist=1.2.0=py39h2bbff1b_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h62dcd97_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py39hc60d5dd_0
h5py=3.6.0=py39h3de5c98_0
hdf5=1.10.6=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.1=h6c2663c_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39haa95532_0
intel-openmp=2021.4.0=haa95532_3556
itsdangerous=2.0.1=pyhd3eb1b0_0
jasper=2.0.33=h77af90b_0
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
keras=2.6.0=pyhd3eb1b0_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libclang=11.1.0=default_h5c34c98_1
libcurl=7.82.0=h86230a5_0
liblapack=3.9.0=5_hd5c7e75_netlib
liblapacke=3.9.0=5_hd5c7e75_netlib
libopencv=4.5.1=py39h27d8466_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libssh2=1.10.0=hcd4344a_0
libtiff=4.2.0=h0c97f57_3
libwebp=1.2.2=h2bbff1b_0
libwebp-base=1.2.2=h8ffe710_1
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=py39haa95532_0
markupsafe=2.0.1=py39h2bbff1b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
msys2-conda-epoch=20160418=1
multidict=5.1.0=py39h2bbff1b_2
numpy=1.21.5=py39h7a0a035_1
numpy-base=1.21.5=py39hca35cd5_1
oauthlib=3.2.0=pyhd3eb1b0_0
opencv=4.5.1=py39hcbf5309_0
openssl=1.1.1n=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
protobuf=3.14.0=py39hd77b12b_1
py-opencv=4.5.1=py39h832f523_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.1.0=py39haa95532_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyreadline=2.1=py39haa95532_1
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-flatbuffers=1.12=pyhd3eb1b0_0
python_abi=3.9=2_cp39
qt=5.12.9=h5909a2a_4
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0.2=py39hf11a4ad_1
scipy=1.7.3=py39h0a974cb_0
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sqlite=3.38.2=h2bbff1b_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py39haa95532_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.6.0=mkl_py39h31650da_0
tensorflow-base=2.6.0=mkl_py39h9201259_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
termcolor=1.1.0=py39haa95532_1
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.35.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
wrapt=1.13.3=py39h2bbff1b_2
xz=5.2.5=h62dcd97_1
yarl=1.6.3=py39h2bbff1b_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.0=h6255e5f_0
